revert the properties extract for jacoco plugin
